% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{ApplyAbundanceFilter}
\alias{ApplyAbundanceFilter}
\title{Function to filter uploaded data}
\usage{
ApplyAbundanceFilter(microSetObj, filt.opt, count, smpl.perc)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function filters data based on low counts in high percentage samples.
Note, first is abundance, followed by variance.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
