% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{ApplyMetaboFilter}
\alias{ApplyMetaboFilter}
\title{Function to filter uploaded metabolomic data}
\usage{
ApplyMetaboFilter(mbSetObj = NA, filter, rsd)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{filtopt}{Character, input the low variance filter option. "iqr" for 
inter-quantile range, "sd" for standard deviation, and "cov" for coefficient of variation.}

\item{filtPerct}{Numeric, input the percentage cutoff for low variance.}
}
\description{
.
Note, this is applied after abundance filter.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
