% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{CalculateHyperScore}
\alias{CalculateHyperScore}
\title{Calculate enrichment score.}
\usage{
CalculateHyperScore(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function calculates the enrichment score for TSEA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
