% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{CompareSummaryStats}
\alias{CompareSummaryStats}
\title{Perform alpha diversity meta-analysis}
\usage{
CompareSummaryStats(
  mbSetObj = NA,
  fileName = "abc",
  sel.meta = "",
  taxrank = "Family",
  view.mode = "ratio",
  format = "png",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{fileName}{File name of output summary image}

\item{sel.meta}{Selected metadata}

\item{taxrank}{Selected taxonomy rank}

\item{view.mode}{Type of plot to output (ratio or box)}

\item{format}{File format of output image (default to png)}

\item{dpi}{dots per inch resolution of image (default 72)}
}
\description{
Compute alpha diversity indices of datasets and plot summary figure, adapted from work of Jordan Bisanz
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
