% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc.R
\name{condenseOTUs}
\alias{condenseOTUs}
\title{Helper Function that will take the OTU Table (as a data.frame())
 to which column representing taxa has been added.
 There is some munging to handle rownames and
 the tax column but the basic idea is there.}
\usage{
condenseOTUs(otutable, splitcol)
}
\description{
Helper Function that will take the OTU Table (as a data.frame())
 to which column representing taxa has been added.
 There is some munging to handle rownames and
 the tax column but the basic idea is there.
}
