% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{CoreMicrobeAnalysis}
\alias{CoreMicrobeAnalysis}
\title{Perform core microbiome analysis}
\usage{
CoreMicrobeAnalysis(mbSetObj, imgName, preval, detection, taxrank, palette,
  viewOpt, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs core microbiome analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
