% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_reporter.R
\name{CreateFooter}
\alias{CreateFooter}
\title{Create report of analyses (Met Enrichment)}
\usage{
CreateFooter(mbSetObj)
}
\description{
Report generation using Sweave
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
