% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{CreatePhyloseqObj}
\alias{CreatePhyloseqObj}
\title{Function to recreate phyloseq object}
\usage{
CreatePhyloseqObj(mbSetObj, type, taxa_type, taxalabel, ismetafile)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function recreates the phyloseq object.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
