% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{CreatePhyloseqObjMeta}
\alias{CreatePhyloseqObjMeta}
\title{Function to recreate phyloseq object in meta-analysis module}
\usage{
CreatePhyloseqObjMeta(
  mbSetObj,
  dataName,
  type,
  taxa_type,
  taxalabel,
  isNormInput
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataName}{Input the file name of selected dataset}

\item{type}{Input the format of the uploaded files. "text" for 
.txt or .csv files, "biom" for .biom, and "mothur" for mothur files.}

\item{taxa_type}{Input the taxonomy labels. "SILVA" for SILVA taxonomy,
"Greengenes" for Greengenes taxonomy, "QIIME" for QIIME taxonomy,
"GreengenesID" for Greengenes OTU IDs, and "Others/Not_specific" for others.}

\item{taxalabel}{Logical, T if taxonomy labels were already included in
the OTU table and F if not.}
}
\description{
This function recreates the phyloseq object.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
