% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{CrossReferencing}
\alias{CrossReferencing}
\title{Perform cross referencing.}
\usage{
CrossReferencing(microSetObj, q.type, module.type)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function performs cross referencing of user's data
with the MicrobiomeAnalyst database. Given a list of species names or ids, 
it finds matched names or ids from selected internal databases.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
