% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{.read.microbiomeanalyst.lib}
\alias{.read.microbiomeanalyst.lib}
\title{Read RDS files from the internet}
\usage{
.read.microbiomeanalyst.lib(filenm, opt = "none")
}
\arguments{
\item{filenm}{Input the name of the file to download}
}
\description{
Function downloads the required file and reads it only if not already in working directory.
Need to specify the file URL and the destfile.
}
