% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{GenerateTemplates}
\alias{GenerateTemplates}
\title{Function to generate templates}
\usage{
GenerateTemplates(mbSetObj, variable)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions generate templates
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
