% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{GetDataForPie}
\alias{GetDataForPie}
\title{Utility function to exract data for piechart.}
\usage{
GetDataForPie(data_n, datataxa, txlvl, OtuIdType, feat_cnt)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function extracts necessary data to create a piechart.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
