% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc.R
\name{GetErrMsg}
\alias{GetErrMsg}
\title{Gets the error message}
\usage{
GetErrMsg()
}
\arguments{
\item{msg}{Error message to print}
}
\description{
The error message will be printed in all cases.
Used in higher functions.
}
