% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{GetGeneListStat}
\alias{GetGeneListStat}
\title{Function to get gene list statistics}
\usage{
GetGeneListStat(microSetObj)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function gets the gene list stats.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
