% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{GetORATable}
\alias{GetORATable}
\title{Getter function}
\usage{
GetORATable(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function retrieves table from mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
