% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{GetSigTable.METAGENOSEQ}
\alias{GetSigTable.METAGENOSEQ}
\title{Getter function}
\usage{
GetSigTable.METAGENOSEQ(microSetObj)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function retrieves table from microSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
