% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{Init.Data}
\alias{Init.Data}
\title{Constructs a microSet object for storing data}
\usage{
Init.Data()
}
\description{
This functions handles the construction of a microSetObj object for storing data 
produced by MicrobiomeAnalystR for further processing and analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
