% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{Init.DataMeta}
\alias{Init.DataMeta}
\title{Init resources & global settings & variables.}
\usage{
Init.DataMeta()
}
\value{
mbSetObj object, Input the name of the mbSetObj. By default,
}
\description{
Init resources & global settings & variables.
}
