% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{Init.mbSetObj}
\alias{Init.mbSetObj}
\title{Constructs a mbSet object for storing data}
\usage{
Init.mbSetObj()
}
\description{
This functions handles the construction of a mbSetObj object for storing data 
produced by MicrobiomeAnalystR for further processing and analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
