% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{MergeDatasets}
\alias{MergeDatasets}
\title{Merge two or more datasets into merged phyloseq object}
\usage{
MergeDatasets(mbSetObj, taxo_type, sample_var)
}
\description{
Merge two or more datasets into merged phyloseq object
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
