% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun_proc.R
\name{PackMeteorResults}
\alias{PackMeteorResults}
\title{Pack merged Meteor2 tables for upload}
\usage{
PackMeteorResults(merged.dir, zip.file, prefix = "output")
}
\arguments{
\item{merged.dir}{Folder with the merged tables.}

\item{zip.file}{Path of the zip file to write.}

\item{prefix}{File prefix used by 'meteor merge' (default "output").}
}
\value{
The path of the zip file (invisibly).
}
\description{
Writes a zip file with the merged species, taxonomy, module, module
completeness and function tables (gene tables are left out) for upload to
MicrobiomeAnalyst.
}
