% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PCoA3D.Anal}
\alias{PCoA3D.Anal}
\title{Main function to perform PCoA analysis}
\usage{
PCoA3D.Anal(
  mbSetObj,
  ordMeth,
  distName,
  taxrank,
  colopt,
  variable,
  taxa,
  alphaopt,
  jsonNm
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{ordMeth}{Character, input the name
of the ordination method. "PCoA" for principal coordinate analysis and "NMDS" for 
non-metric multidimensional scaling.}

\item{distName}{Character, input the name of the distance method.}

\item{taxrank}{Character, input the taxonomic
level for beta-diversity analysis.}

\item{colopt}{Character, color the data points by the experimental factor,
the taxon abundance of a selected taxa, or alpha diversity.}

\item{variable}{Character, input the name of the experimental factor.}

\item{taxa}{Character, if the data points are colored by taxon abundance, 
input the name of the selected taxa.}

\item{alphaopt}{Character, if the data points are colored by alpha-diversity, 
input the preferred alpha-diversity measure.}

\item{jsonNm}{Character, input the name of the json file to output.}

\item{datatype}{Character, input "16S" if the data is marker
gene data and "metageno" if it is metagenomic data.}
}
\description{
This functions creates a 3D PCoA plot from the microbiome data.
This is used by the Beta-Diversity analysis.
The 3D interactive visualization is on the web.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
