% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_utils.R
\name{PCoA3DAnal.16SRef}
\alias{PCoA3DAnal.16SRef}
\title{Function to create PCoA}
\usage{
PCoA3DAnal.16SRef(microSetObj, barplotNm, ordMeth, distName, datatype,
  taxrank, colopt, taxa, alphaopt, metadata, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function creates data for plotting PCoA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
