% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{Perform16FunAnot_mem}
\alias{Perform16FunAnot_mem}
\title{Main function to perform 16S functional annotation}
\usage{
Perform16FunAnot_mem(mbSetObj, type, pipeline, ggversion)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This is the main function to perform either PICRUSt or
SILVA for functional annotation on the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
