% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_qc.R
\name{PerformBatchCorrection}
\alias{PerformBatchCorrection}
\title{Perform batch correction of meta-analysis datasets}
\usage{
PerformBatchCorrection()
}
\description{
Perform batch correction of meta-analysis datasets
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
