% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PerformKOEnrichAnalysis_KO01100}
\alias{PerformKOEnrichAnalysis_KO01100}
\title{Function to prepare KO enrichment analysis.}
\usage{
PerformKOEnrichAnalysis_KO01100(mbSetObj, category, file.nm)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function performs KO enrichment analysis
using the KO01100 map.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
