% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PerformKOmapping}
\alias{PerformKOmapping}
\title{Function to perform KO mapping.}
\usage{
PerformKOmapping(microSetObj, geneIDs, type)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function performs KO mapping.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
