% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PerformKOProjection}
\alias{PerformKOProjection}
\title{Function to perform KO projection}
\usage{
PerformKOProjection(microSetObj)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function projects user-uploaded
KOs onto the KEGG Global Metabolic Network.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
