% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformLefseAnal}
\alias{PerformLefseAnal}
\title{Main function to perform LEfSe analysis}
\usage{
PerformLefseAnal(mbSetObj, p.lvl, lda.lvl, variable, isfunc, datatype,
  shotgunid, taxrank)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs LEfSe analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
