% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PerformMetaboNormalization}
\alias{PerformMetaboNormalization}
\title{Function to perform normalization on metabolomics data}
\usage{
PerformMetaboNormalization(
  mbSetObj,
  rowNorm,
  transNorm,
  scaleNorm,
  isAutoScale
)
}
\arguments{
\item{transNorm}{Select option to transform the data, "LogNorm" for Log Normalization,
and "CrNorm" for Cubic Root Transformation.}

\item{scaleNorm}{Select option for scaling the data, "MeanCenter" for Mean Centering,
"AutoNorm" for Autoscaling, "ParetoNorm" for Pareto Scaling, amd "RangeNorm" for Range Scaling.}

\item{ref}{Input the name of the reference sample or the reference feature, use " " around the name.}

\item{ratio}{This option is only for biomarker analysis.}

\item{ratioNum}{Relevant only for biomarker analysis.}
}
\description{
This function performs normalization on the uploaded metabolomics
data.
"SumNorm" for Normalization to constant sum, 
"MedianNorm" for Normalization to sample median, and
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
