% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PerformMetaEffectSize}
\alias{PerformMetaEffectSize}
\title{Perform differential abundance analysis on individual dataset and apply combine effect size meta-analysis method on the results}
\usage{
PerformMetaEffectSize(
  mbSetObj = NA,
  imgName = "",
  taxrank = "OTU",
  selMeta,
  BHth = 0.05,
  de.method = "LM",
  ef.method = "REML",
  format = "png",
  dpi = 72
)
}
\arguments{
\item{imgName}{File name of output bar plot image}

\item{selMeta}{selected metadata}

\item{BHth}{P-value threshold}

\item{de.method}{Differential abundance method}

\item{ef.method}{Effect size model}

\item{format}{file format of output image}

\item{dpi}{dots per inch resolution of image (default 72)}

\item{method}{Method used, either "rem" or "fem"}

\item{taxRank}{selected taxa rank which to perform analysis}
}
\description{
Perform differential abundance analysis on individual dataset and apply combine effect size meta-analysis method on the results
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
