% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformMetagenomeSeqAnal}
\alias{PerformMetagenomeSeqAnal}
\title{Main function to perform metagenome seq analysis}
\usage{
PerformMetagenomeSeqAnal(mbSetObj, variable, p.lvl, datatype, shotgunid,
  taxrank, model)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs metagenome seq analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
