% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PerformNetworkCorrelation}
\alias{PerformNetworkCorrelation}
\title{Function to call for correlation network}
\usage{
PerformNetworkCorrelation(
  mbSetObj,
  taxrank,
  cor.method = "pearson",
  colorOpt = "expr",
  permNum = 100,
  pvalCutoff = 0.05,
  corrCutoff = 0.3,
  abundOpt = "mean",
  corr.net.name
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function runs the fastspar or cor.test
}
