% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PerformPartialCorr}
\alias{PerformPartialCorr}
\title{Perform Partial Correlation Analysis}
\usage{
PerformPartialCorr(mbSetObj, taxa.lvl = "Phylum", variable = NA,
  alg = "pearson", pval.cutoff = 0.05, imgName = "partial_corr",
  format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{taxa.lvl}{Character, input the taxonomic level
to perform partial correlation analysis.}

\item{variable}{Character, input the selected variable.}

\item{alg}{Use "kendall" or "spearman" for non-parametric and 
"pearson" for parametric.}

\item{imgName}{Character, input the name of the partial correlation plot.}

\item{format}{Character, by default the plot format
is "png".}

\item{dpi}{Numeric, set to 72 by default.}
}
\description{
Function to perform and plot
partial correlations between all taxonomic features,
the outcome, and selected confounders.
NOTE: All metadata must be numeric
}
