% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_utils.R
\name{PerformRefDataMapping}
\alias{PerformRefDataMapping}
\title{Function to perform reference data mapping.}
\usage{
PerformRefDataMapping(mbSetObj, refdataNm, taxo_type, sample_var, biome)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function performs reference data mapping.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
