% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformUnivarTest}
\alias{PerformUnivarTest}
\title{Main function to perform univariate analysis.}
\usage{
PerformUnivarTest(mbSetObj, variable, p.lvl, datatype, shotgunid, taxrank,
  statOpt)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs univariate analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
