% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotAlphaBoxData}
\alias{PlotAlphaBoxData}
\title{Function to create box plots for alpha diversity analysis}
\usage{
PlotAlphaBoxData(mbSetObj, boxplotName, distName, metadata,
  colors = "default", format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{boxplotName}{Character, input the name of the boxplot.}

\item{distName}{Character, input the diversity measure
to calculate alpha-diversity. "Chao1", "Observed", "ACE", "Shannon", 
"Simpson", or "Fisher".}

\item{metadata}{Input the name of the experimental factor to group the samples.}

\item{colors}{Use "default", "viridis" to 
use the viridis color palette, "plasma" to use the plasma
color palette, "magma" to use the magma color palette, or
"magma" to use the magma color palette.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions performs metagenome seq analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
