% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotAlphaData}
\alias{PlotAlphaData}
\title{Function to plot alpha-diversity analysis.}
\usage{
PlotAlphaData(mbSetObj, data.src, bargraphName, distName, metadata,
  taxrank, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions creates a plot for alpha-diversity.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
