% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_methods.R
\name{PlotBetaSummary}
\alias{PlotBetaSummary}
\title{Perform beta diversity meta-analysis}
\usage{
PlotBetaSummary(
  mbSetObj,
  plotNm,
  taxalvl,
  sel.meta,
  alg,
  format = "png",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{plotNm}{File name of output summary image}

\item{taxalvl}{Selected taxonomy rank}

\item{sel.meta}{Selected metadata}

\item{alg}{statistical comparison algorithm to detect significant different between metadata groups}

\item{format}{File format of output image (default to png)}

\item{dpi}{dots per inch resolution of image (default 72)}
}
\description{
Compute beta diversity indices of datasets and plot summary figure, adapted from work of Jordan Bisanz
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
