% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotBoxDataCorr}
\alias{PlotBoxDataCorr}
\title{Function to create box plots of important features}
\usage{
PlotBoxDataCorr(
  mbSetObj,
  boxplotName,
  feat,
  format = "png",
  sel.meta = "",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{boxplotName}{Character, input the name of the 
box plot.}

\item{feat}{Character, input the name of the selected 
feature.}

\item{format}{Character, by default the plot format
is "png".}

\item{dpi}{Dots per inch. Numeric, by default
it is set to 72.}
}
\description{
This functions plots box plots of a selected feature.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
