% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotCorr}
\alias{PlotCorr}
\title{Main function to plot correlation matrix}
\usage{
PlotCorr(mbSetObj, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots the correlation matrix.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
