% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotCorrHeatMap}
\alias{PlotCorrHeatMap}
\title{Function to create correlation heat map}
\usage{
PlotCorrHeatMap(mbSetObj, imgName, format = "png", width = NA,
  cor.method, colors_cntrst, viewOpt, taxrank, fix.col, no.clst, top,
  topNum, datatype)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function creates the correlation heat map
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
