% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{PlotEnrichNet.Overview}
\alias{PlotEnrichNet.Overview}
\title{Create network for enrichmnet overview}
\usage{
PlotEnrichNet.Overview(hits, pvals)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function creates the plot
for the enrichent network overview.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
