% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotFunAnotSummary}
\alias{PlotFunAnotSummary}
\title{Plot functional annotation summary}
\usage{
PlotFunAnotSummary(microSetObj, imgName, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions plots the functional annotation summary.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
