% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Main function to plot heatmap.}
\usage{
PlotHeatmap(mbSetObj, plotNm, smplDist, clstDist, palette, metadata,
  taxrank, datatype, viewOpt, doclust, format = "png", showfeatname,
  appendnm, rowV = F, colV = T, var.inx = NA, border = T,
  width = NA, dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots a heatmap from the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
