% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotImpVar}
\alias{PlotImpVar}
\title{Helper function to plot variable importance}
\usage{
PlotImpVar(microSetObj, imp.vec, xlbl, feat.num = 15, color.BW = FALSE)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions plots the variable importance
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
