% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotImpVarLEfSe}
\alias{PlotImpVarLEfSe}
\title{Plot LEfSe summary}
\usage{
PlotImpVarLEfSe(mbSetObj, imp.vec, layoutOptlf, meta, colOpt = "default",
  color.BW = FALSE)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions graphically summarizes the LEfSe results
using a bargraph.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
