% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotLEfSeSummary}
\alias{PlotLEfSeSummary}
\title{Plot LEfSe summary}
\usage{
PlotLEfSeSummary(mbSetObj, ldaFeature, imgName, format = "png",
  width = NA, dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{ldaFeature}{Numeric, input the number of top
features to include in the plot.}

\item{imgName}{Character, input the name
of the plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{width}{Numeric, input the width of the plot. By
default it is set to NA.}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions graphically summarizes the LEfSe results
in a VIP plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
