% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PlotLibSizeView}
\alias{PlotLibSizeView}
\title{Function to plot library size}
\usage{
PlotLibSizeView(mbSetObj, imgName, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function creates a plot summarizing the library
size of microbiome dataset.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
