% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_qc.R
\name{PlotMetaDensity}
\alias{PlotMetaDensity}
\title{Plot density plot of datasets in meta-analysis module}
\usage{
PlotMetaDensity(imgNm, dpi = 72, format = "png", factor = "")
}
\arguments{
\item{imgNm}{name of the image to output}

\item{dpi}{dots per inch resolution of image (default 72)}

\item{format}{File format of output image (default to png)}
}
\description{
Plot density plot of datasets in meta-analysis module
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
