% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotNetworkSummary}
\alias{PlotNetworkSummary}
\title{Function to create Network Summary plot}
\usage{
PlotNetworkSummary(
  mbSetObj,
  imgName,
  taxa,
  format,
  width = NA,
  dpi = 72,
  colOpt = "default",
  num_taxa = 10
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{colOpt}{Character, "default", "viridis",
"cividis", or "plasma".}
}
\description{
This function outputs the network
summary plot for a specific taxa.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
