% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotPartialCorrConfounders}
\alias{PlotPartialCorrConfounders}
\title{Function to perform and plot
partial correlations between a selected feature,
the outcome, and selected confounders.
NOTE: All metadata must be numeric}
\usage{
PlotPartialCorrConfounders(mbSetObj, pcorrOpt = "default", feat = "NA",
  test = "lefse", p.cutoff = 0.05, lda.cutoff = 2, fc.cutoff = 2,
  variable = NA, alg = "pearson", colOpt = "D",
  imgName = "partial_corr", format = "png", feats.cutoff = 15)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{pcorrOpt}{Character, either use the top 15 features 
or perform partial correlation using a single feature.}

\item{feat}{Character, input the name of the specific feature
to perform partial correlation.}

\item{test}{Character, input the analysis used to determine
the set of important features.}

\item{p.cutoff}{Numeric.}

\item{lda.cutoff}{Numeric.}

\item{fc.cutoff}{Numeric.}

\item{variable}{Character.}

\item{alg}{Use "kendall" or "spearman" for non-parametric and 
"pearson" for parametric.}

\item{colOpt}{Five options are available. "magma" (or "A"), "inferno" (or "B"), 
"plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}

\item{imgName}{Character, input the name of the partial correlation plot.}

\item{format}{Character, by default the plot format
is "png".}

\item{feats.cutoff}{Numeric, insert the maximum number of important features
to include. Default is set to 15.}
}
\description{
Function to perform and plot
partial correlations between a selected feature,
the outcome, and selected confounders.
NOTE: All metadata must be numeric
}
