% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotPartialCorrelationConfounders}
\alias{PlotPartialCorrelationConfounders}
\title{Function to perform and plot
partial correlations between a selected feature,
the outcome, and selected confounders.
NOTE: All metadata must be numeric}
\usage{
PlotPartialCorrelationConfounders(mbSetObj, test = "lefse",
  p.cutoff = 0.05, lda.cutoff = 2, fc.cutoff = 2, variable = NA,
  alg = "pearson", colOpt = "D", imgName = "partial_corr",
  format = "png", confounder1 = NA, confounder2 = NA,
  confounder3 = NA, confounders = NA)
}
\arguments{
\item{test}{Character, input the analysis used to determine
the set of important features.}

\item{alg}{Use "kendall" or "spearman" for non-parametric and 
"pearson" for parametric.}

\item{colOpt}{Five options are available. "magma" (or "A"), "inferno" (or "B"), 
"plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E")}

\item{confounders}{Vector containing list of confounders
to control for.}
}
\description{
Function to perform and plot
partial correlations between a selected feature,
the outcome, and selected confounders.
NOTE: All metadata must be numeric
}
