% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PlotPCAView}
\alias{PlotPCAView}
\title{Plot PCA plot for multi-omics samples}
\usage{
PlotPCAView(imgName, format = "png", dpi = 72, init)
}
\arguments{
\item{imgNm}{name of the image to output}
}
\description{
Plot PCA plot for multi-omics samples
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
