% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotPhylogeneticTree}
\alias{PlotPhylogeneticTree}
\title{Function to prepare data for phylogenetic tree.}
\usage{
PlotPhylogeneticTree(mbSetObj, color, shape, taxa, treeshape, imgName,
  format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions prepares the data to plot the phylogenetic tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
