% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PlotRareCurve}
\alias{PlotRareCurve}
\title{Function to plot rarefraction curves}
\usage{
PlotRareCurve(mbSetObj, graphName, variable)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function plots rarefraction curves from the uploaded
data for both sample-wise or group-wise
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
