% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotRNASeqDotPlot}
\alias{PlotRNASeqDotPlot}
\title{Plot fold-change dot plot}
\usage{
PlotRNASeqDotPlot(mbSetObj, top.inx = 15, colpal = "plasma",
  imgName = "rnaseq_dot", format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{top.inx}{Numeric, input the number of features to include in the dot plot.}

\item{colpal}{Character, "default", "viridis" for the 
default viridis color palette, "plasma"
for the plasma color palette, and "cividis" for the  
cividis color palette.}

\item{imgName}{Character, input the name of the plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions creates a dot plot of the RNAseq analysis 
results from the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
