% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotRNAseqVolcano}
\alias{PlotRNAseqVolcano}
\title{Function to plot volacano analysis results from RNAseq analysis}
\usage{
PlotRNAseqVolcano(mbSetObj, fc.cutoff = 2, p.cutoff = 0.05,
  colpal = "default", imgName = "volcano", format = "png",
  dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{fc.cutoff}{Numeric, input the fold-change cutoff.}

\item{p.cutoff}{Numeric, input the p-value cutoff.}

\item{colpal}{Character, "default" for a blue and red color
palette, "pog" for the purple and orange color palette, "bog"
for the blue and orange color palette, and "rgg" for the red and 
green color palette.}

\item{imgName}{Character, input the name of the plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions plots the RNAseq analysis 
results from the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
