% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotTaxaAbundanceBarSamGrp}
\alias{PlotTaxaAbundanceBarSamGrp}
\title{Function to plot group-wise bar charts.}
\usage{
PlotTaxaAbundanceBarSamGrp(
  mbSetObj,
  barplotName,
  taxalvl,
  metadata,
  facet2,
  imgOpt,
  feat_cnt,
  colpalopt,
  calcmeth,
  toptaxa,
  abunTopTaxaOpt,
  appendnm,
  format = "png",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots group-wise bar charts of a specified 
taxa for alpha diversity analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
