% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotTaxaAlphaArea}
\alias{PlotTaxaAlphaArea}
\title{Plot functional annotation summary}
\usage{
PlotTaxaAlphaArea(microSetObj, barplotName, viewOpt, taxalvl, metadata,
  imgOpt, feat_cnt, colpalopt, calcmeth, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions plots the functional annotation summary.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
