% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotTaxaAlphaBar}
\alias{PlotTaxaAlphaBar}
\title{Function to plot bar charts for alpha diversity}
\usage{
PlotTaxaAlphaBar(mbSetObj, barplotName, taxalvl, metadata, facet, imgOpt,
  feat_cnt, colpalopt, calcmeth, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots bar charts of different taxonomic levels
for alpha diversity.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
