% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{PrepareEnrichNet}
\alias{PrepareEnrichNet}
\title{Function to prepare data for enrichment network.}
\usage{
PrepareEnrichNet(microSetObj)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function prepares data for enrichment network.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
