% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PrepareHeatTreePlot}
\alias{PrepareHeatTreePlot}
\title{Function to plot heat tree}
\usage{
PrepareHeatTreePlot(mbSetObj, meta, tax, color, layoutOpt, comparison,
  imgName, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots the heat tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
