% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PrepareHeatTreePlotDataParse_cmf}
\alias{PrepareHeatTreePlotDataParse_cmf}
\title{Function to prepare heat tree data}
\usage{
PrepareHeatTreePlotDataParse_cmf(dm_otu_cmf, dm_samples_cmf, meta)
}
\arguments{
\item{meta}{Input the name of the group.}
}
\description{
This functions plots the heat tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
