% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PrepareHeatTreePlotDataParse_cmf_diff_table}
\alias{PrepareHeatTreePlotDataParse_cmf_diff_table}
\title{Function to prepare heat tree data}
\usage{

  PrepareHeatTreePlotDataParse_cmf_diff_table(PrepareHeatTreePlotDataParse_cmf_res)
}
\description{
This functions plots the heat tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
