% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_utils.R
\name{PrepareMergedData}
\alias{PrepareMergedData}
\title{Function to merge user data with public data.}
\usage{
PrepareMergedData(microSetObj, metadata, keepfeat)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function is used to merge microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
