% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PreparePCA4Shotgun}
\alias{PreparePCA4Shotgun}
\title{Function to prepare shotgun data for PCA.}
\usage{
PreparePCA4Shotgun(
  mbSetObj,
  imgName,
  imgName2,
  format = "json",
  inx1,
  inx2,
  inx3,
  variable,
  showlabel,
  format2d = "png",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function formats shotgun data for PCA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
