% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_reporter.R
\name{PreparePDFReport}
\alias{PreparePDFReport}
\title{Function to create PDF report}
\usage{
PreparePDFReport(microSetObj, usrName)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function creates a PDF report.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
