% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PrepareQueryJson}
\alias{PrepareQueryJson}
\title{Function to prepare query for JSON.}
\usage{
PrepareQueryJson(microSetObj)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function prepares the data for JSON.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
