% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{Read16SAbundData}
\alias{Read16SAbundData}
\title{Main function to read 16S data}
\usage{
Read16SAbundData(mbSetObj, dataName, type, taxalabel, taxa_type,
  ismetafile, module.type)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This is the main function read in the 16S data
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
