% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{Read16SAbundDataMeta}
\alias{Read16SAbundDataMeta}
\title{Function to read 16S Abundance data in meta-analysis module}
\usage{
Read16SAbundDataMeta(
  mbSetObj,
  dataName,
  format,
  taxa_type,
  ismetafile,
  is.normalized = F
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataName}{File name of selected dataset}

\item{format}{File type, only supports ".txt"}

\item{taxa_type}{Taxonomy type (i.e.}

\item{ismetafile}{Whether meta-data is given (for BIOM format only)}

\item{is.normalized}{Specify whether uploaded data is normalized or raw counts}
}
\description{
This is the main function read in the 16S data
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
