% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{Read16STaxaTable}
\alias{Read16STaxaTable}
\title{Function to read 16S taxonomy table from txt format}
\usage{
Read16STaxaTable(microSetObj, dataName)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function reads in the 16S taxonomy table from txt format
into the microSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
