% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{Read16STaxaTableMeta}
\alias{Read16STaxaTableMeta}
\title{Function to read 16S taxonomy table from txt format in meta-analysis module}
\usage{
Read16STaxaTableMeta(mbSetObj, fileName, dataName)
}
\arguments{
\item{mbSetObj}{Input the name of selected dataset's taxa file}

\item{dataName}{Input the name of selected dataset data file}
}
\description{
This function reads in the 16S taxonomy table from txt format
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
