% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{ReadMetabolicTable}
\alias{ReadMetabolicTable}
\title{Main function to read metabolomics table}
\usage{
ReadMetabolicTable(mbSetObj, dataName, metType, idType)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{metType}{whether a peak or metabolite intensity table}

\item{idType}{One of compound name,KEGG or HMDB ID. Only applicable when the metType is metabolite. 
McGill University, Canada
License: GNU GPL (>= 2)}
}
\description{
This is the main function read in the metabolomics data
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
}
