% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{ReadMothurData}
\alias{ReadMothurData}
\title{Function to read 16S data in mothur format}
\usage{
ReadMothurData(mbSetObj, dataName, taxdataNm, taxa_type)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function reads in the 16S data from mothur format
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
