% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{ReadSampleTableMeta}
\alias{ReadSampleTableMeta}
\title{Function to read in sample data in meta-analysis module (merged metadata file)}
\usage{
ReadSampleTableMeta(mbSetObj, fileName)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{fileName}{Input the name of merged metadata file}
}
\description{
This functions reads in sample data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
