% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{ReadSampleTableMetaInd}
\alias{ReadSampleTableMetaInd}
\title{Function to read in sample data in meta-analysis module}
\usage{
ReadSampleTableMetaInd(mbSetObj, dataName, metaName)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataName}{Input the file name of selected dataset}

\item{metaName}{Input the file name of meta-data file name.}
}
\description{
This functions reads in sample data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
