% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{ReadShotgunBiomData}
\alias{ReadShotgunBiomData}
\title{Main function to read in shotgun data}
\usage{
ReadShotgunBiomData(mbSetObj, dataName, geneidtype, module.type,
  ismetadata)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function reads in shotgun data
in biom format.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
