% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{ReadTreeFile}
\alias{ReadTreeFile}
\title{Function to read in tree files.}
\usage{
ReadTreeFile(mbSetObj, dataName)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataName}{Input the tree file name.}
}
\description{
This functions reads in tree files.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
