% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{RF.Anal}
\alias{RF.Anal}
\title{Performs Random Forest Analysis}
\usage{
RF.Anal(microSetObj, treeNum, tryNum, randomOn, variable, taxrank,
  datatype)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs the random forest analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
