% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{SanityCheckData}
\alias{SanityCheckData}
\title{Main function to sanity check on uploaded microbiome data}
\usage{
SanityCheckData(mbSetObj, filetype, disableFilter = FALSE)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{filetype}{Character, "biom" if the uploaded data
was .biom format, "mothur" if mothur format, or "txt" if
as .txt or .csv format.}

\item{disableFilter}{Boolean. Set to TRUE to bypass the hard-coded filter
to remove features occuring in <2 samples.}
}
\description{
This function performs a sanity check on the uploaded
user data. It checks the grouping of samples, if a phylogenetic tree
was uploaded.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
