% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{SanityCheckDataMeta}
\alias{SanityCheckDataMeta}
\title{Individual sanity check on selected datase in meta-analysis module}
\usage{
SanityCheckDataMeta(mbSetObj, dataName, filetype, disableFilter = FALSE)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataName}{Input the name of selected dataset data file}

\item{filetype}{Character, "biom" if the uploaded data
was .biom format, "mothur" if mothur format, or "txt" if
as .txt or .csv format.}

\item{disableFilter}{Boolean. Set to TRUE to bypass the hard-coded filter
to remove features occuring in <2 samples.}
}
\description{
This function performs a sanity check on the uploaded
user data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
