% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_metadata.R
\name{SanityCheckSampleData}
\alias{SanityCheckSampleData}
\title{SanityCheckSampleData}
\usage{
SanityCheckSampleData(mbSetObj = NA)
}
\arguments{
\item{mbSetObj}{microbiomeanalyst object, initialized by InitDataObjects("pktable", "mf", FALSE)}
}
\description{
SanityCheckSampleData
}
