% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{SanityCheckSampleDataMeta}
\alias{SanityCheckSampleDataMeta}
\title{SanityCheckSampleData}
\usage{
SanityCheckSampleDataMeta(mbSetObj = NA, dataName = "")
}
\arguments{
\item{mbSetObj}{microbiomeanalyst object, initialized by InitDataObjects("pktable", "mf", FALSE)}

\item{dataName}{File name of selected dataset}
}
\description{
SanityCheckSampleData
}
