% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{SetAnalType}
\alias{SetAnalType}
\title{Function to set analysis type}
\usage{
SetAnalType(analType)
}
\arguments{
\item{analType}{Input the analysis type. If the data is marker gene data, 
use "markergene", if the data is shotgun metagenomics or transcriptomics data, 
use ""shotgun". If performing the Projection with Public Data module, use "dataprojection". 
If performing Taxon Set Enrichment Analysis, use "species".}
}
\description{
This functions sets the analysis mode.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
