% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_metadata.R
\name{SetDataTypeOfMeta}
\alias{SetDataTypeOfMeta}
\title{SetDataTypeOfMeta}
\usage{
SetDataTypeOfMeta(mbSetObj = NA)
}
\arguments{
\item{mbSetObj}{microbiomeanalyst object}
}
\description{
SetDataTypeOfMeta
}
