% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifac_metadata.R
\name{SetMetaAttributes}
\alias{SetMetaAttributes}
\title{SetMetaAttributes'}
\usage{
SetMetaAttributes(mbSetObj = NA, init = 1)
}
\arguments{
\item{mbSetObj}{microbiomeanalyst object}

\item{init}{can be 0 or 1}
}
\description{
SetMetaAttributes'
}
