% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{SetMsetLib}
\alias{SetMsetLib}
\title{Set the microbe set library}
\usage{
SetMsetLib(microSetObj, tset.type)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function sets the microbe
set library for TSEA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
