% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_multivar.R
\name{UpdateConfItems}
\alias{UpdateConfItems}
\title{Function to update confounders used for partial correlation}
\usage{
UpdateConfItems(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function updates which confounders will be
used to calculate partial correlation.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
