% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{UpdateListInput}
\alias{UpdateListInput}
\title{Function to filter list data based on a minimum count}
\usage{
UpdateListInput(microSetObj, minL, maxL = Inf)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function filters the inputted list data
based on a minimum count.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
