% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{UpdatePieData}
\alias{UpdatePieData}
\title{Function to update pie-chart data.}
\usage{
UpdatePieData(microSetObj, lowtaxa)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions updates pie chart data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
