% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{ApplyVarianceFilter}
\alias{ApplyVarianceFilter}
\title{Function to filter uploaded data}
\usage{
ApplyVarianceFilter(microSetObj, filtopt, filtPerct)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function filters data based on low abundace or variance.
Note, this is applied after abundance filter.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
