% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{CoreMicrobeAnalysis}
\alias{CoreMicrobeAnalysis}
\title{Perform core microbiome analysis}
\usage{
CoreMicrobeAnalysis(
  mbSetObj,
  imgName,
  preval,
  detection,
  taxrank,
  palette,
  viewOpt,
  analOpt,
  expFact,
  group,
  format = "png",
  dpi = 72,
  width = NA
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{imgName}{Input the name of the core microbiome analysis plot.}

\item{preval}{Numeric, input the sample prevalence cutoff. This is
the minimum number of samples that share this microbe.}

\item{detection}{Numeric, input the relative abundance cutoff. This
is the minimum detection threshold of X percent relative abundance.}

\item{palette}{Character, input the color palette.
"bwm" for the default color palette /(blue, white, red/), "gbr"
for the red and green color palette, "heat" for the yellow to red color
palette, "topo" for the blue, green, and yellow color palette, 
"gray" for the gray color palette, "byr" for the blue, yellow and red color
palette, "viridis" for the viridis default color palette and "plasma"
for the plasma viridis color palette.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions performs core microbiome analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
