% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_reporter.R
\name{CreateSDPRnwReport}
\alias{CreateSDPRnwReport}
\title{Function to create PDF report for SDP module}
\usage{
CreateSDPRnwReport(microSetObj, usrName)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This function creates a PDF report
for the SDP module - writes .Rnw file template.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
