% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{FeatureCorrelation}
\alias{FeatureCorrelation}
\title{Main function to calculate correlation of all other feature to a given feature name}
\usage{
FeatureCorrelation(mbSetObj, dist.name, taxrank, feat)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dist.name}{Input the distance measure. Input "pearson" for 
Pearson R, "spearman" for Spearman rank correlation, and "kendall"
for Kendall rank correlation.}

\item{taxrank}{Character, input the taxonomic level to use.}

\item{taxa}{If the pattern is defined by using a specific taxon, input the name
of that taxa here.}

\item{variable}{Input the name of the experimental factor.}

\item{datatype}{If the data is marker gene, input "16S". If the
data is metagenomic, use "metageno".}

\item{shotfeat}{Only valid for SDP module, set to "null".}

\item{shotgunid}{Only valid for SDP module, set to "NA".}
}
\description{
This functions calculate correlation of all other feature to a given feature name.
This is used in the Pattern Search analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
