% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{FeatureCorrelation}
\alias{FeatureCorrelation}
\title{Main function to calculate correlation of all other feature to a given feature name}
\usage{
FeatureCorrelation(mbSetObj, dist.name, taxrank, taxa, variable, datatype,
  shotfeat, shotgunid)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions calculate correlation of all other feature to a given feature name.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
