% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{GetColorSchema}
\alias{GetColorSchema}
\title{Function to get color palette for graphics.}
\usage{
GetColorSchema(mbSetObj, grayscale = F)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{grayscale}{Logical, default set to F.}
}
\description{
This function is called to create a color palette
based on the number of groups. It returns a vector of color
hex codes based on the number of groups.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
