% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{GetFinalNameMap}
\alias{GetFinalNameMap}
\title{Getter to return final map}
\usage{
GetFinalNameMap(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function returns the final (after user selection) map as a dataframe.
Consists of two columns, original name and strain.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
