% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{GetMapTable}
\alias{GetMapTable}
\title{Getter function}
\usage{
GetMapTable(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function retrieves table from mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
