% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{Match.Pattern}
\alias{Match.Pattern}
\title{Function to match patterns}
\usage{
Match.Pattern(mbSetObj, dist.name = "pearson", pattern = NULL, taxrank,
  taxa, variable, datatype, shotfeat, shotgunid)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions matches patterns
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
