% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PCoA3D.Anal}
\alias{PCoA3D.Anal}
\title{Main function to perform PCoA analysis}
\usage{
PCoA3D.Anal(microSetObj, ordMeth, distName, datatype, taxrank, colopt,
  variable, taxa, alphaopt, jsonNm)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs PCoA analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
