% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_utils.R
\name{PCoA3DAnal.16SRef}
\alias{PCoA3DAnal.16SRef}
\title{Function to create PCoA}
\usage{
PCoA3DAnal.16SRef(mbSetObj, barplotNm, ordMeth, distName, taxrank,
  metadata, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function creates data for plotting PCoA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
