% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PerformAlphaDiversityComp}
\alias{PerformAlphaDiversityComp}
\title{Perform alpha diversity}
\usage{
PerformAlphaDiversityComp(microSetObj, opt, metadata)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs alpha diversity.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
