% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PerformCategoryComp}
\alias{PerformCategoryComp}
\title{Function to perform categorical comparison.}
\usage{
PerformCategoryComp(mbSetObj, taxaLvl, method, distnm, variable)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{taxaLvl}{Input the name of the taxonomic level to calculate Beta-diversity comparison.}

\item{method}{Statistical method to calculate
beta-diversity significance. Use "adonis" for Permutational MANOVA, "anosim" for
Analysis of Group Similarities and "permdisp" for
Homogeneity of Group Dispersions.}

\item{distnm}{Character, input the name of the distance method. "bray" for
Bray-Curtis Index, "jsd" for Jensen-Shannon Divergence, "jaccard" for Jaccard Index,
"unifrac" for Unweighted Unifrac Distance and "wunifrac" for Weighted Unifrac Distance.}

\item{variable}{Input the name of the experimental factor to group the samples.}
}
\description{
This functions performs categorical comparisons.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
