% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PerformCategoryComp}
\alias{PerformCategoryComp}
\title{Function to perform categorical comparison.}
\usage{
PerformCategoryComp(mbSetObj, method, distnm, variable)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs categorical comparisons.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
