% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PerformKOEnrichAnalysis_Table}
\alias{PerformKOEnrichAnalysis_Table}
\title{Perform KO Enrichment Analysis}
\usage{
PerformKOEnrichAnalysis_Table(microSetObj, file.nm)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs KO enrichment analysis
on a tabled input.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
