% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformMetagenomeSeqAnal}
\alias{PerformMetagenomeSeqAnal}
\title{Main function to perform metagenome seq analysis}
\usage{
PerformMetagenomeSeqAnal(mbSetObj, variable, p.lvl, datatype, shotgunid,
  taxrank, model)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{variable}{Character, input the name of the experimental factor.}

\item{p.lvl}{Numeric, input the adjusted p-value cutoff.}

\item{datatype}{Character, input whether the data is marker gene
data ("16S") or metagenomic data ("metageno").}

\item{shotgunid}{Only valid for SDP module, set to "NA".}

\item{taxrank}{Character, input the taxonomic level to perform
univariate analysis.}

\item{model}{Character, input the name of the statistical
model to fit the data. Use "zigfit" for zero-inflated Gaussian fit
and "ffm" for the fitFeatureModel.}
}
\description{
This functions performs metagenome seq analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
