% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PerformNormalization}
\alias{PerformNormalization}
\title{Function to perform normalization}
\usage{
PerformNormalization(mbSetObj, rare.opt, scale.opt, transform.opt)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function performs normalization on the uploaded
data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
