% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PerformNormalization}
\alias{PerformNormalization}
\title{Function to perform normalization}
\usage{
PerformNormalization(mbSetObj, rare.opt, scale.opt, transform.opt)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{rare.opt}{Character, "rarewi" to rarefy the data and 
"none" to not.}

\item{scale.opt}{Character, input the name of the data scaling option.
"colsum" for total sum scaling, "CSS" for cumulative sum scaling,
"upperquartile" for upper-quartile normalization, and "none" to none.}

\item{transform.opt}{Character, input the name of the data transformation
to be applied. "rle" for relative log expression, "TMM" for trimmed mean of 
M-values, "clr" for centered log ratio, and "none" for none.}
}
\description{
This function performs normalization on the uploaded
data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
