% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{PerformRarefaction}
\alias{PerformRarefaction}
\title{Utility function to perform rarefraction (used by PerformNormalization)}
\usage{
PerformRarefaction(mbSetObj, data, rare.opt)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function performs rarefraction on the uploaded
data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
