% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformRNAseqDE}
\alias{PerformRNAseqDE}
\title{Main function to perform RNAseq analysis}
\usage{
PerformRNAseqDE(microSetObj, opts, p.lvl, variable, datatype, shotgunid,
  taxrank)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs RNAseq analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
