% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PerformUnivarTest}
\alias{PerformUnivarTest}
\title{Main function to perform univariate analysis.}
\usage{
PerformUnivarTest(microSetObj, variable, p.lvl, datatype, shotgunid,
  taxrank, statOpt)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions performs univariate analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
