% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotAlphaBoxData}
\alias{PlotAlphaBoxData}
\title{Function to create box plots for alpha diversity analysis}
\usage{
PlotAlphaBoxData(mbSetObj, boxplotName, distName, metadata,
  format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions performs metagenome seq analysis on the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
