% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotBetaDiversity}
\alias{PlotBetaDiversity}
\title{Function to plot beta diversity.}
\usage{
PlotBetaDiversity(mbSetObj, plotNm, ordmeth, distName, colopt, metadata,
  showlabel, taxrank, taxa, alphaopt, ellopt, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions creates beta diversity plots.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
