% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotBoxData}
\alias{PlotBoxData}
\title{Function to create box plots of important features}
\usage{
PlotBoxData(mbSetObj, boxplotName, feat, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots box plots of a selected feature.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
