% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotCorrHeatMap}
\alias{PlotCorrHeatMap}
\title{Function to create correlation heat map}
\usage{
PlotCorrHeatMap(mbSetObj, imgName, format = "png", width = NA,
  cor.method, colors_cntrst, viewOpt, taxrank, fix.col, no.clst, top,
  topNum, datatype)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{imgName}{Character, input the name
of the plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{width}{Numeric, input the width of the plot. By
default it is set to NA.}

\item{cor.method}{Input the distance measure. Input "pearson" for 
Pearson R, "spearman" for Spearman rank correlation, and "kendall"
for Kendall rank correlation.}

\item{colors_cntrst}{Set the colors of the heatmap. By default it 
is set to "bwm", blue, white, to red. Use "gbr" for green, black, red, use
"heat" for red to yellow, "topo" for blue to yellow, "gray" for 
white to black, and "byr" for blue, yellow, red. Use "viridis"
for the viridis color palette and "plasma" for the plasma color palette from
the viridis R package.}

\item{viewOpt}{Character, "overview" to view an overview
of the heatmap, and "detail" to iew a detailed view of the
heatmap (< 1500 features).}

\item{taxrank}{Character, input the taxonomic level to perform
classification. For instance, "OTU-level" to use OTUs.}

\item{fix.col}{Logical, default set to FALSE.}

\item{no.clst}{Logical, default set to FALSE.}

\item{top}{Logical, default set to FALSE.}

\item{topNum}{Numeric, set the number of top features
to include.}

\item{datatype}{If the data is marker gene, input "16S". If the
data is metagenomic, use "metageno".}
}
\description{
This function creates the correlation heat map
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
