% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotDataPieFromPie}
\alias{PlotDataPieFromPie}
\title{Function to plot pie-chart data.}
\usage{
PlotDataPieFromPie(mbSetObj, taxalvl, metadata, clslevel, taxaposn,
  lowtaxa)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots pie charts of microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
