% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{PlotFunctionStack}
\alias{PlotFunctionStack}
\title{Function to plot stacked bar chart of functional data.}
\usage{
PlotFunctionStack(mbSetObj, summaryplot, functionlvl, abundcal, geneidtype,
  metadata, colpalopt, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function plots stacked bar charts of functional data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
