% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotGroupPieGraph}
\alias{PlotGroupPieGraph}
\title{Main function to plot pie graphs of microbiome data.}
\usage{
PlotGroupPieGraph(mbSetObj, taxalvl, metadata, clslevel, feat_cnt,
  calcmeth)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots pie graphs of microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
