% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Main function to plot heatmap.}
\usage{
PlotHeatmap(
  mbSetObj,
  plotNm,
  dataOpt = "norm",
  scaleOpt = "row",
  smplDist,
  clstDist,
  palette,
  metadata,
  taxrank,
  viewOpt,
  doclust,
  format = "png",
  colname,
  rowname,
  fontsize_col,
  fontsize_row,
  annoPer,
  fzAnno,
  appendnm = "F",
  rowV = F,
  colV = T,
  var.inx = NA,
  border = T,
  width = NA,
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{plotNm}{Character, input the name
of the plot.}

\item{smplDist}{Input the distance measure. "euclidean" for
Euclidean distance, "correlation" for Pearson, and "minkowski"
for Minkowski.}

\item{clstDist}{Character, input the name of the
selected clustering algorithm. "ward" for Ward, "average" for Average, 
"complete" for Complete, and "single" for Single.}

\item{palette}{Set the colors of the heatmap. By default it 
is set to "bwm", blue, white, to red. Use "gbr" for green, black, red, use
"heat" for red to yellow, "topo" for blue to yellow, "gray" for 
white to black, and "byr" for blue, yellow, red.}

\item{metadata}{Character, input the name of the experimental factor 
to cluster samples by.}

\item{taxrank}{Character, input the taxonomic level to perform
classification. For instance, "OTU-level" to use OTUs.}

\item{viewOpt}{Character, "overview" to view an overview
of the heatmap, and "detail" to iew a detailed view of the
heatmap (< 1500 features).}

\item{doclust}{Logicial, default set to "F".}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{colname}{Logical, specify the if column name is shown, default set to "T",}

\item{rowname}{Logical, specify the if row name is shown, default set to "T",}

\item{fontsize_col}{Numeric, fontsize for colnames}

\item{fontsize_row}{Numeric, fontsize for rownames}

\item{appendnm}{Logical, "T" to prepend higher taxon names.}

\item{rowV}{Logical, default set to "F".}

\item{colV}{Logical, default set to "T".}

\item{var.inx}{Default set to NA.}

\item{width}{Numeric, input the width of the plot. By
default it is set to NA.}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions plots a heatmap from the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
