% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotLEfSeSummary}
\alias{PlotLEfSeSummary}
\title{Plot LEfSe summary}
\usage{
PlotLEfSeSummary(microSetObj, imgName, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions graphically summarizes the LEfSe results
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
