% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotLEfSeSummary}
\alias{PlotLEfSeSummary}
\title{Plot LEfSe summary}
\usage{
PlotLEfSeSummary(mbSetObj, ldaFeature, imgName, format = "png",
  width = NA, dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions graphically summarizes the LEfSe results
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
