% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotOverallPieGraph}
\alias{PlotOverallPieGraph}
\title{Main function to plot pie graphs of microbiome data.}
\usage{
PlotOverallPieGraph(microSetObj, taxalvl, feat_cnt, calcmeth)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions plots pie graphs of microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
