% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotPiechart}
\alias{PlotPiechart}
\title{Function to create pie-chart plot.}
\usage{
PlotPiechart(microSetObj, rel_perct, pieName, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions creates the pie chart plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
