% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotRarefactionCurve}
\alias{PlotRarefactionCurve}
\title{Function to create rarefraction curves of microbiome data}
\usage{
PlotRarefactionCurve(microSetObj, data.src, linecolor, linetype, facet,
  step, imgName, format = "png", dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions plots rarefraction curves.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
