% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotRF.Classify}
\alias{PlotRF.Classify}
\title{Plot random forest classification}
\usage{
PlotRF.Classify(mbSetObj, feature, imgName, format = "png", dpi = 72,
  width = NA)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots the classification of samples
from the random forest analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
