% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{PlotRF.VIP}
\alias{PlotRF.VIP}
\title{Plot variable importance ranked by MeanDecreaseAccuracy}
\usage{
PlotRF.VIP(mbSetObj, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plot variable importance ranked by MeanDecreaseAccuracy
from the random forest analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
