% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotSamplePieGraph}
\alias{PlotSamplePieGraph}
\title{Main function to plot sample-wise pie graphs of microbiome data.}
\usage{
PlotSamplePieGraph(
  mbSetObj,
  taxalvl,
  smplnm,
  feat_cnt,
  toptaxapie,
  pietoptaxaopt
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{taxalvl}{Character, input the taxonomic level to perform
classification. For instance, "Genus" to use the Genus level.}

\item{feat_cnt}{Set the minimum feature count that is used to "bin"
together small taxa.}
}
\description{
This functions plots sample-wise pie graphs of microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
