% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PlotTaxaAlphaBarSamGrp}
\alias{PlotTaxaAlphaBarSamGrp}
\title{Function to plot group-wise bar charts.}
\usage{
PlotTaxaAlphaBarSamGrp(mbSetObj, barplotName, taxalvl, metadata, imgOpt,
  feat_cnt, colpalopt, calcmeth, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions plots group-wise bar charts of a specified 
taxa for alpha diversity analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
