% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotTreeGraph}
\alias{PlotTreeGraph}
\title{Main function to plot tree graphics}
\usage{
PlotTreeGraph(microSetObj, plotNm, distnm, clstDist, metadata, datatype,
  taxrank, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions creates tree plots from the microbiome data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
