% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_utils.R
\name{PlotUsrRefPCoA3DScore}
\alias{PlotUsrRefPCoA3DScore}
\title{Function to plot 3D score plot.}
\usage{
PlotUsrRefPCoA3DScore(mbSetObj, imgName, format = "json", inx1, inx2,
  inx3, variable)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function plots a 3D PCoA score plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
