% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PrepareHeatTreePlot}
\alias{PrepareHeatTreePlot}
\title{Function to plot heat tree}
\usage{
PrepareHeatTreePlot(
  mbSetObj,
  meta,
  taxalvl,
  color,
  layoutOpt,
  comparison,
  wilcox.cutoff,
  imgName,
  format = "png",
  dpi = 72
)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{meta}{Input the name of the group.}

\item{taxalvl}{Character, input the taxonomic level to perform
classification. For instance, "Genus" to use the Genus level.}

\item{color}{Character, input the color palette code. "dbgr"
for dark blue, grey and red. "bgy" for greenblue, grey and yellow.
"ggr" for green, grey and red. "pgy" for purple, gray and yellow. 
"tgr" for teal, grey and red. "ggg" for green, grey and gold. Additionally,
the viridis R package can be used to generate color schemes. "plasma" for 
the plasma color scheme, "viridis" for the default viridis color scheme and
"cividis" for the cividis color scheme.}

\item{layoutOpt}{Character, input the layout of the heat tree. "dft"
for the default layout and "reda" for reingold-tilford.}

\item{comparison}{Character, input the group comparisons.}

\item{wilcox.cutoff}{Numeric, input the Wilcoxan p-value cutoff
for significant node labels.}

\item{imgName}{Character, input the name of the heat tree plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions plots the heat tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
