% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{PreparePhylogeneticTreePlot}
\alias{PreparePhylogeneticTreePlot}
\title{Function to prepare data for phylogenetic tree.}
\usage{
PreparePhylogeneticTreePlot(mbSetObj, color, shape, taxa, treeshape,
  imgName, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions prepares the data to plot the phylogenetic tree.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
