% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{Read16SBiomData}
\alias{Read16SBiomData}
\title{Function to read 16S data in biom format}
\usage{
Read16SBiomData(mbSetObj, dataName, taxa_type, ismetadata)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function reads in the 16S data from biom format
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
