% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_io.R
\name{Read16STabData}
\alias{Read16STabData}
\title{Function to read 16S data}
\usage{
Read16STabData(mbSetObj, dataName)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This is the main function read in the 16S data in txt format
into the mbSetObj.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
