% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_io.R
\name{ReadSampleTable}
\alias{ReadSampleTable}
\title{Function to read in sample data}
\usage{
ReadSampleTable(mbSetObj, dataName)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This functions reads in sample data.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
