% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utils.R
\name{ReadShotgunTabData}
\alias{ReadShotgunTabData}
\title{Main function to read in shotgun data}
\usage{
ReadShotgunTabData(mbSetObj, dataName, geneidtype, datatype)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function reads in shotgun data
in a tabular format.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
