% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_anal.R
\name{RunFastSpar}
\alias{RunFastSpar}
\title{Function to call FastSpar}
\usage{
RunFastSpar(mbSetObj, taxrank, permNum, pvalCutoff, corrCutoff,
  output = "network", palette = "viridis")
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function runs the fastspar
}
