% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{SanityCheckData}
\alias{SanityCheckData}
\title{Main function to sanity check on uploaded data}
\usage{
SanityCheckData(mbSetObj, datatype, filetype)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function performs a sanity check on the uploaded
user data. It checks the grouping of samples, if a phylogenetic tree
was uploaded.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
