% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{SavePiechartImg}
\alias{SavePiechartImg}
\title{Function to save pie-chart}
\usage{
SavePiechartImg(mbSetObj, taxalvl, pieName, format = "png", dpi = 72)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{taxalvl}{Character, input the taxonomic level to perform
classification. For instance, "Genus" to use the Genus level.}

\item{pieName}{Character, input the name of the pie chart plot.}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}
}
\description{
This functions saves created pie chart plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
