% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_utils.R
\name{SavePiechartImg}
\alias{SavePiechartImg}
\title{Function to save pie-chart}
\usage{
SavePiechartImg(microSetObj, taxalvl, pieName, format = "png",
  dpi = 72)
}
\arguments{
\item{microSetObj}{Input the name of the microSetObj.}
}
\description{
This functions saves created pie chart plot.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
