% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_utils.R
\name{Setup.MapData}
\alias{Setup.MapData}
\title{Function to set up data for TSEA}
\usage{
Setup.MapData(mbSetObj, qvec)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function sets up data for TSEA.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
