% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_proc.R
\name{UpdateSampleItems}
\alias{UpdateSampleItems}
\title{Function to update samples}
\usage{
UpdateSampleItems(mbSetObj)
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}
}
\description{
This function prunes samples to be included 
for downstream analysis.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
